// types/biologicalOccurences.ts
export type BiologicalOccurencesType = {
    sudorese: boolean
    bruxismo: boolean
    apneiaDoSono: boolean
    ronco: boolean
    movimentosPeriodicosDosMembros: boolean
    despertaresParciais: boolean
    refluxoGastroesofagico: boolean
    sialorreia: boolean
    arritmias: boolean
    mioclonia: boolean
    parassonia: boolean
    epistaxe: boolean
    miccaoInvoluntaria: boolean
    evacuacaoInvoluntaria: boolean
    polucao: boolean
}

export const DefaultBiologicalOccurences: BiologicalOccurencesType = {
    sudorese: false,
    bruxismo: false,
    apneiaDoSono: false,
    ronco: false,
    movimentosPeriodicosDosMembros: false,
    despertaresParciais: false,
    refluxoGastroesofagico: false,
    sialorreia: false,
    arritmias: false,
    mioclonia: false,
    parassonia: false,
    epistaxe: false,
    miccaoInvoluntaria: false,
    evacuacaoInvoluntaria: false,
    polucao: false,
}